+++ app/src/main/java/de/triology/universeadm/DisableCacheResponseFilter.java
@@ -1,40 +1,32 @@
 package de.triology.universeadm;
 
-//~--- JDK imports ------------------------------------------------------------
-
+import javax.servlet.Filter;
+import javax.servlet.FilterChain;
+import javax.servlet.FilterConfig;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
+import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-import javax.ws.rs.container.ContainerRequestContext;
-import javax.ws.rs.container.ContainerResponseContext;
-import javax.ws.rs.container.ContainerResponseFilter;
-import javax.ws.rs.ext.Provider;
-
-/**
- *
- * @author Sebastian Sdorra <sebastian.sdorra@triology.de>
- */
-@Provider
-public class DisableCacheResponseFilter implements ContainerResponseFilter
+public class DisableCacheResponseFilter implements Filter
 {
+    @Override
+    public void init(FilterConfig filterConfig) throws ServletException {}
 
-  /**
-   * Method description
-   *
-   *
-   * @param requestContext
-   * @param responseContext
-   *
-   * @throws IOException
-   */
-  @Override
-  public void filter(ContainerRequestContext requestContext,
-    ContainerResponseContext responseContext)
-    throws IOException
-  {
-    responseContext.getHeaders().putSingle("Cache-Control",
-      "no-cache, no-store, must-revalidate");
-    responseContext.getHeaders().putSingle("Pragma", "no-cache");
-    responseContext.getHeaders().putSingle("Expires", 0);
-    
-  }
+    @Override
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
+        throws IOException, ServletException
+    {
+        HttpServletResponse httpResponse = (HttpServletResponse) response;
+
+        httpResponse.setHeader("Cache-Control", "no-cache, no-store, must-revalidate");
+        httpResponse.setHeader("Pragma", "no-cache");
+        httpResponse.setDateHeader("Expires", 0);
+
+        chain.doFilter(request, response);
+    }
+
+    @Override
+    public void destroy() {}
 }