+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#238] Fix a bug where caching prevented the cas redirect flow.
+  - Convert the disable cache filter to a servlet filter and use it for every request and not only for jax-rs endpoints.
 
 ## [v1.21.0-4] - 2026-04-17
 ### Added